move data to api module
